--- src/components/Cotizador.styled.js
@@ -1,13 +1,0 @@
-import styled from 'styled-components';
-//import { rgba } from 'polished';
-//import { media } from 'styled-bootstrap-grid';
-
-//import { colors } from '@bn3/ui/lib/theme';
-
-const Page = styled.div`
-    fontFamily: 'Helvetica';
-    backgroundColor: '#f0f0f0';
-    flexDirection: 'column';
-  `;
-
-export { Page };
